Refactored - Removed unnecessary actions - Moved actions
